Added firebase remote config with default channel

+++ app/src/main/java/com/allrecipes/managers/GoogleYoutubeApiManager.java
@@ -23,7 +23,7 @@ public class GoogleYoutubeApiManager {
     public Observable<SearchChannelVideosResponse> fetchChannelVideos(
         String channelId,
         String pageToken,
-        int maxResults,
+        long maxResults,
         String order,
         String searchCriteria
     ) {
